+++ web/src/search/input/QueryBuilderInputRow.tsx
@@ -26,9 +26,11 @@ interface Props {
     /** A list of useful examples demonstrating valid values that can be inputted for this search field. */
     examples: QueryFieldExamples[]
     /** Handler for when an input field changes. */
-    onInputChange: (
-        key: keyof QueryBuilderState['fields']
-    ) => (event: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) => void
+    onInputChange: Record<
+        keyof QueryBuilderState['fields'],
+        React.ChangeEventHandler<HTMLInputElement | HTMLSelectElement>
+    >
+
     isSourcegraphDotCom: boolean
 }
 
@@ -48,7 +50,7 @@ export const QueryBuilderInputRow: React.FunctionComponent<Props> = props => {
                     autoCapitalize="off"
                     autoComplete="off"
                     placeholder={placeholder}
-                    onChange={props.onInputChange(props.shortName)}
+                    onChange={props.onInputChange[props.shortName]}
                 />
             </div>
             <InfoDropdown title={props.title} markdown={props.description} examples={props.examples} />