+++ CHANGELOG.md
@@ -15,7 +15,17 @@ All notable changes to Sourcegraph are documented in this file.
 
 ### Removed
 
-## 3.1
+## 3.1.1
+
+### Added
+
+- Query builder toggle (open/closed) state is now retained.
+
+### Fixed
+
+- Fixed an issue where single-term values entered into the "Exact match" field in the query builder were not getting wrapped in quotes.
+
+## 3.1.0
 
 ### Added
 
@@ -28,8 +38,8 @@ All notable changes to Sourcegraph are documented in this file.
 
 ### Changed
 
-- File match search results now show full repo name if there are results from mirrors on different code hosts (e.g. github.com/sourcegraph/sourcegraph and gitlab.com/sourcegraph/sourcegraph)
-- Search queries now use "smart case" by default. Searches are case insensitive unless you use uppercase letters. To explicitely set the case, you can still use the `case` field (e.g. `case:yes`, `case:no`). To explicitely set smart case, use `case:auto`.
+- File match search results now show full repository name if there are results from mirrors on different code hosts (e.g. github.com/sourcegraph/sourcegraph and gitlab.com/sourcegraph/sourcegraph)
+- Search queries now use "smart case" by default. Searches are case insensitive unless you use uppercase letters. To explicitly set the case, you can still use the `case` field (e.g. `case:yes`, `case:no`). To explicitly set smart case, use `case:auto`.
 
 ### Fixed